requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
pillow==12.0.0
pip==25.3
pyside6==6.10.0
pyside6-addons==6.10.0
pyside6-essentials==6.10.0
shiboken6==6.10.0
